Remove `assets` group from Gemfile
Rails 4 removes this group so we can remove it and include the gems that
were in it for all groups. In Rails 4 the production environment will
not try to dynamically generate any assets by default and use the static
precompiled assets and to go along with this change the assets are now
intended to be precompiled in the production environment.

If you're running rake assets:compile to precompile the assets you'll
need to set the environment to production first.

    RAILS_ENV=production bundle exec rake assets:precompile

http://railscasts.com/episodes/415-upgrading-to-rails-4?view=asciicast