Remove Turbolinks
The line `//= require turbolinks` in `application.js` is causing
Suspenders to throw a `Sprockets::FileNotFound couldn't find file
'turbolinks'` error.

This most likely happened when Suspenders was upgraded to support
Rails 4 where Turbolinks are the default. There are two possible
solutions to the problem: either to remove this line or to install the
`turbolinks` gem.

We're choosing the former for now until we've fully evaluated Turbolinks
on real projects.