Run `rake dev:prime` on CI
Previously, the `dev:prime` task was only available in the development
environment, while CI runs in the test environment. Because the environments
didn't match, the `dev:prime` task is never defined in CI, and CI crashes.

We now define the `dev:prime` task in the test environment, because CI should be
testing whether this part of `bin/setup` runs too.