Set auto-deploy for Heroku suspended apps
Previously, the Circle CI config for suspended applications enabled
automatic deployments for all apps. Given that the `bin/deploy` script
is only generated for apps suspended with the `--heroku true` flag, this
led to builds failing on the deployment step due to the script not being
found.

This PR adds the deployment command to `circle.yml` only when an app is
suspended with the Heroku options. There are still manual setup steps
required, such as setting the Heroku API and SSH key but those are
clearly indicated by the error messages on Circle CI and makes
Suspenders defaults more "out-of-the-box" friendly.

[fixes #610]