Provide Shoulda Matchers config
The recent release of Shoulda Matchers 3.0 introduces some changes,
namely requiring configuration in order to operate.
The absence of the file leads to hard to decipher errors: they indicate
methods are missing and can lead to some frustration trying to
understand what is necessary to get Shoulda to behave correctly.

This PR introduces a build step to provide a config file for in order
for it to be enabled "out-of-the-box". It opts in to enable interaction
with all of Rails, as this seems the principle of least surprise.