Don't load i18n tasks in production
This gem is used in development and test and is not required in
production. At minimum, loading it in production is a memory use issue
(however small), but it likely also adds overhead to every translation
to track used keys.

This could explain some complaints we've seen of slow view rendering on
apps that I know to have been recently generated by suspenders.