Remove sprocket-es6
Why:

sprockets-es6 is causing depreciation issues for Sprockets, and the gem
author says that the gem will never reach v1.0.

This PR:

Removes sprockets-es6. We would like to introduce the babel-transpiler
gem in the future but it requires sprockets 4. While sprockets 4 has a
beta release, this seems to be causing issues with bourbon neat and
bitter per https://github.com/thoughtbot/suspenders/issues/797. Until
this is resolved I suggest adding ES6 support on a project by project
basis.