Introduce JavaScript driver generator (#882)
* Introduce JavaScript driver generator

This introduces `suspenders:js_driver`, a generator for installing our
JavaScript test runner. This one is capybara-webkit.

---

While here, remove the `configure_ci` build step that was undefined in a
prior commit.

* No need to bundle in the JS driver generator

We only need to run `bundle install` from the generator when we want to
then run further generators or otherwise make use of the newly-added
gem. We do no such thing here.