Move enforce_ssl to production:force_tls
Move the `suspenders:enforce_ssl` generator to
`suspenders:production:force_tls`. This namespaces it as something
specific to production (useful for those who are looking for the
Heroku-specific functionality), and it renames SSL to TLS, as that is
the abbreviation used in this century.

While here, shuffle some things around, which exposes existing bugs.
And fix those bugs!