suspenders:production:email can control app.json
The `SMTP_ADDRESS`, `SMTP_DOMAIN`, `SMTP_PASSWORD`, and `SMTP_USERNAME`
variables are only required when dealing with emails in production, so
reflect that in our `app.json` and the `suspenders:production:email`
generator.

This introduces our first test for destroying a generator. It also
introduces some awesome code for un-merging a Hash from another Hash.

The Heroku actions create the `app.json`, so make sure they run before
we modify it. It makes sense to run all production generators after the
Heroku-specific actions, so make that clumping happen.