Relax the newline requirements
Look, a newline here or there to break up the text can go a long way
toward improving readability. Let's embrace that.

That said, a newline at the beginning or an extra one at the end is too
much. Plus, too many in the middle is clearly a sign of things gone
wrong.