Drop jquery-rails gem
Rails dropped jQuery as a dependency back in
https://github.com/rails/rails/pull/27113. That commit removed
'jquery-ujs' from the application.js and replaced it with 'rails-ujs'.
Since we aren't using 'jquery-ujs', we don't need the 'jquery-rails'
gem.

As for the test JS, `$.fx.off = true` shouldn't be necessary if
we aren't using jQuery, and `$.ajaxSetup({ async: false });` seems
to have been for Akephalos, which we no longer use.