Introduce suspenders:preloader
This introduces the `suspenders:preloader` generator, for managing
Spring. This generator will completely remove Spring or completely add
it, much in the same way `rails new --skip-spring` would.

Rails uses `--skip-spring` (or the absence of) to determine four things:

1. [Install][install-config] a [default `Spring.watch` configuration].
2. [Add the Spring and spring-watcher-listen gems][gems].
3. Do not [cache classes] in test runs.
4. Run [`spring binstub --all`][spring-binstub].

[install-config]: https://github.com/rails/rails/blob/41bbd7cd0e4e950ea5a366f78b26b32ed6bd0644/railties/lib/rails/generators/rails/app/app_generator.rb#L112
[default `Spring.watch` configuration]: https://github.com/rails/rails/blob/master/railties/lib/rails/generators/rails/app/templates/config/spring.rb.tt
[spring-binstub]: https://github.com/rails/rails/blob/2a3f759eef10352bedce5f13b12dbdda30aacab2/railties/lib/rails/generators/app_base.rb#L424-L426
[gems]: https://github.com/rails/rails/blob/64ddabe58eb492f852f01fbe1cc0712264b63ecc/railties/lib/rails/generators/rails/app/templates/Gemfile.tt#L58-L64
[cache classes]: https://github.com/rails/rails/blob/f73093bcdf6a63ea7a47743fd8f94f6cab4380f9/railties/lib/rails/generators/rails/app/templates/config/environments/test.rb.tt#L12

The `spring binstub --all` command will add Spring to any file
registered as desiring a Spring binstub. This can be controlled by
third-party gems, such as [spring-commands-rspec] which adds it to the
`bin/rspec` file. We should keep an eye out for Spring integration for
any new binstubs as we add them to our `bin/` directory.

[spring-commands-rspec]: https://github.com/jonleighton/spring-commands-rspec

We had previously run `spring binstub --all` earlier in the Suspenders
chain, but I moved it to the end. This means that Spring will
potentially generate binstubs for more commands.

The `Spring.watch` configuration is duplicated into Suspenders so we can
add it after the fact, as people desire.

---

In terms of implementation, we want to run some Thor actions
(`gsub_file` and `run`) during both `invoke!` and `revoke!` -- that is,
during both `rails generate` and `rails destroy`. However, those actions
have [hard-coded checks] that prevent them from running in a `revoke!`. We
solve this by duplicating them into our code, deleting the check, and
modifying as needed.

[hard-coded checks]: https://github.com/erikhuda/thor/commit/528a51f1b001953d9272a65a55571270ce856c1f

In the future we will need to watch for [Rails subtly using
non-reversable Thor actions][rails-pr-35285].

[rails-pr-35285]: https://github.com/rails/rails/pull/35285

The `spring` gem can come as part of the Gemfile as generated by Rails,
which means it will be in a `group :development` block. The `gem` Rails
generator action only works with non-block gem requirements in the
Gemfile. Use the `gem_group` Rails generator action instead.

A few unfortunate ideas have collided to make the `gem`/`gem_group`
actions unpleasant:

- The Gemfile contains any valid Ruby, so parsing it and generating it
is as hard as parsing and generating Ruby.
- `gem "foo", group: :bar` and `group(:bar) { gem "foo" }` are
semantically equivalent.
- The `gem` and `gem_group` actions work on text, not on semantics.
- Bundler does not provide helpers for manipulating, parsing, or
generating Gemfiles.

Ideally we would work at the object level, not the raw text level --
_or_ we would work at the text level and not the object level. Having
two semantically equivalent mechanisms for specifying the group makes us
want a more clear mechanism for programmatically editing the Gemfile.