Remove .ctags configuration
Not all of us are blessed with ctags-compatible editors these days, but
that's not the main thing.

Those of us who are using an editor built for programming are [running
ctags automatically], and therefore the configuration in `.ctags` is not
useful.

[running ctags automatically]: https://github.com/ludovicchabant/vim-gutentags/

Of the projects I found on my laptop, none have modified `.ctags` beyond
this default.

This file made more sense [when it was added], but now it's a relic of a
different era.

[when it was added]: https://github.com/thoughtbot/suspenders/pull/595