Improve reliability and aesthetics of comments stripper (#1085)
Reliably remove comments from Rails config files when generating
Suspenders apps. With the parser gem, we're now removing comments by
processing their surrounding code using AST (abstract syntax tree)
processors.

We were using a regex to do that, which is text-based and thereby
subject to edge cases that only a proper AST solution can handle.

** First motivation

This pull request, https://github.com/thoughtbot/suspenders/pull/1044,
made changes to some config files to fix the Action Mailer asset host.

While working on merging that PR, the tests were not passing because
they looked for an output that was missing in one of the changed
config files:

    config.action_mailer.asset_host = \
      %q{"https://#{ENV\.fetch\("ASSET_HOST", ENV\.fetch\("APPLICATION_HOST"\)\)}}

The PR code changes are kind of straightforward and that config option
should have been present in the config file but it was not.

The reason was that Suspenders was removing code comments with an
overly permissive regex : (/^.*#.*$/), which removed every line
containing a # character, no matter where they were in the line - so
it was also removing the new Action Mailer config setting.

That regex was also silently messing up with other config lines, for
example:

    config.public_file_server.headers = {
      'Cache-Control' => "public, max-age=#{2.days.to_i}"
    }

As a quick fix to merge that PR, I've changed the regex to /^\s*#.*$/
but it was not an ideal solution - which is what I'm striving for with
this commit. With the old text-based regex solution, there were still
nasty edge cases to deal with, and we may never know when they'd
occur. Also, the previous regex did not work with inline comments.

** Second motivation

The final aesthetics of the config files could be improved in my
opinion. This was how they looked like:

    Rails.application.configure do
      config.cache_classes = false
      config.eager_load = false
      config.consider_all_requests_local = true
      if Rails.root.join('tmp', 'caching-dev.txt').exist?
        config.action_controller.perform_caching = true
        config.action_controller.enable_fragment_cache_logging = true
        config.cache_store = :memory_store
        config.public_file_server.headers = {
          'Cache-Control' => "public, max-age=#{2.days.to_i}"
        }
      else
        config.action_controller.perform_caching = false
        config.cache_store = :null_store
      end
      config.active_storage.service = :local
      config.action_mailer.raise_delivery_errors = true
      config.after_initialize do
        Bullet.enable = true
        Bullet.bullet_logger = true
        Bullet.rails_logger = true
      end

      # ...
    end

And with the newest solution they look like this:

    Rails.application.configure do
      config.cache_classes = false
      config.eager_load = false
      config.consider_all_requests_local = true

      if Rails.root.join('tmp', 'caching-dev.txt').exist?
        config.action_controller.perform_caching = true
        config.action_controller.enable_fragment_cache_logging = true

        config.cache_store = :memory_store
        config.public_file_server.headers = {
          'Cache-Control' => "public, max-age=#{2.days.to_i}"
        }
      else
        config.action_controller.perform_caching = false
        config.cache_store = :null_store
      end

      config.action_mailer.raise_delivery_errors = true

      config.after_initialize do
        Bullet.enable = true
        Bullet.bullet_logger = true
        Bullet.rails_logger = true
      end

      # ...

Which preserves newlines where appropriate, so that's easier to read.

** Third motivation

This pull request uses the parser gem, and I feel like introducing
that gem in the project may be a good move over the long term.

For example, we need to finish up the gem action to sort gem
declarations alphabetically when adding gems to the Gemfile, and the
most reliable way to do that is with AST parsing.

** Other related changes

1. Make the indentation of some templates match the Rails
environment's config files'. This goes along with the aesthetics
introduced in this commit.

2. Because of (1), I've also found a bug that was concatenating two
lines of codes in config/environments/production.rb with no blank line
between them (in email_generator.rb), so I'm taking the opportunity to
fix it

3. standardrb fixes after updating to the latest standardrb version